third_party/WebKit/LayoutTests/editing/execCommand/insert-list-infinite-loop2.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
  '<table contenteditable="true"><tbody><tr><td>',
    '<table><tbody><tr><td>',
      '<ol>',
        '^<li>AAA<br><br></li><li><br></li><li>BBB<br>CCC</li>|',
      '</ol>',
    '</td></tr></tbody></table>',
    'DDD',
    '<div style="-webkit-user-select: none;"></div>',
    '<div style="display: inline-block; border:solid;"></div>',
  '</td></tr></tbody></table>'
  ],
  'insertOrderedList',
  [
  '<table contenteditable="true"><tbody><tr><td>',
    '<ol>',
      '<li>',
        '<table><tbody><tr><td>',
          '^AAA<br><br>BBB<br>CCC',
        '</td></tr></tbody></table>|',
      '</li>',
    '</ol>',
    'DDD',
    '<div style="-webkit-user-select: none;"></div>',
    '<div style="display: inline-block; border:solid;"></div>',
  '</td></tr></tbody></table>'
  ]),
  'InsertList should work inside an empty div with no placeholder, without any infinite loop');
</script>
